// apps/api/src/visit-tracking/enums/channel-type.enum.ts
export enum ChannelType {
  ORGANIC_SEARCH = 'organic_search',
  PAID_SEARCH = 'paid_search',
  SOCIAL_ORGANIC = 'social_organic',
  SOCIAL_PAID = 'social_paid',
  REFERRAL = 'referral',
  OWNED_REFERRAL = 'owned_referral',
  INTERNAL = 'internal',
  DIRECT = 'direct',
  EMAIL = 'email',
  AFFILIATE = 'affiliate',
  OTHER = 'other',
}

export enum DeviceType {
  MOBILE = 'mobile',
  TABLET = 'tablet',
  DESKTOP = 'desktop',
}